Extract remote project tool executor
