+++ docs/DECISIONS.md
@@ -105,6 +105,7 @@
 - Sidebar command presentation has one source of truth in `QuillCodeSidebarCommandPresentation`. It owns visible primary/sidebar utility command ordering, display labels, native SF Symbols, HTML icon tokens, and Playwright test IDs so the SwiftUI sidebar and static HTML harness cannot drift as the Codex-like rail evolves.
 - Project row actions mirror thread row actions: the rail stays scannable, and New chat, Refresh context, Rename, and Remove from list live behind an ellipsis menu plus matching command-palette and slash routes. Removing a project forgets the project registry entry and detaches chats from that workspace root, but it never deletes files or thread transcripts.
 - Project connections are typed metadata, not path string conventions. Local projects use `Local`; registered SSH projects use `SSH Remote` everywhere in sidebar badges, command-palette titles, slash suggestions, and disabled-action explanations. SSH Remote rows can be registered with `/ssh user@host:/path` or `Project: Add SSH Remote...`. The first remote executor is noninteractive SSH: integrated-terminal commands, agent-authored `host.shell.run`, bounded `host.file.read`/`host.file.write`, `host.apply_patch`, `host.git.status`/`host.git.diff`, file/hunk `host.git.stage`/`host.git.restore`, `host.git.commit`, `host.git.push`, `host.git.pr.create`/`view`/`checks`/`checkout`/`comment`/`review`/`reviewers`/`labels`/`merge`, and `host.git.worktree.*` calls run through `ssh -T -o BatchMode=yes` from the selected remote root. SSH Remote projects intentionally offer the agent only remote-safe base tools (`host.shell.run`, bounded file read/write, apply-patch, git status/diff/stage/restore/commit/push/PR/worktree, plus app/plan/browser-style tools); environment tools are withheld until remote adapters exist, and unexpected local-only tool calls fail in the transcript instead of falling back to the desktop workspace.
+- SSH Remote project tool execution lives in `WorkspaceRemoteProjectToolExecutor.swift`. The workspace model owns selected-project orchestration and transcript side effects, while the executor owns the remote-safe tool catalog, agent override construction, command construction, path normalization, artifact labeling, and unsupported-tool errors. Future QuillCloud relay or alternate remote transports should add an adapter behind this boundary instead of adding remote command branches back to `WorkspaceModel.swift`.
 - SSH Remote context refresh is a read-only SSH probe, not a sync layer. It reads bounded `AGENTS.md`, `.quillcode/rules.md`, `.quillcode/instructions.md`, and project `.quillcode/memories` files from the remote root into the normal project/thread context models, using hex-encoded records so paths and content with spaces survive the shell transport. Direct remote file read/write is limited to normalized project-relative paths through the SSH executor; remote apply-patch reuses the same unsafe diff-path scan and then runs `git apply --check` before `git apply` remotely. Remote hunk stage/restore reuses the git patch path-mismatch scan and `git apply --check` before applying over SSH. Remote commit trims and validates commit messages before running `git commit -m`; remote push reuses the local git-name validator for explicit refs and validates discovered current branches before pushing; remote PR creation delegates to the remote GitHub CLI after validating base/head refs; remote PR read/check/checkout/comment/review/reviewer-update/label-update/merge operations delegate to `gh pr view`, `gh pr checks`, `gh pr checkout`, `gh pr comment`, `gh pr review`, `gh pr edit`, and `gh pr merge` after validating selectors, checkout branch names, review actions, reviewer names, label names, merge methods, and body requirements; remote worktree create/remove uses the same sibling-of-project path boundary and verifies removal targets with remote `git worktree list --porcelain`. Plugin and local-environment actions still need explicit remote adapters rather than local fallbacks.
 - Execution context is shared surface metadata, not renderer guesswork. Project-bound tool cards and terminal history entries carry an optional `ExecutionContextSurface`; local contexts stay visually muted, while SSH Remote contexts render a compact chip plus a thin left rail in SwiftUI, static HTML, and the Playwright harness. This keeps old transcript scrollback clear even after the user switches projects and leaves room for future QuillCloud/relay contexts to reuse the same contract.
 - Tool-card execution-context enrichment lives in `WorkspaceExecutionContextSurfaceBuilder`. The workspace model supplies selected-project and project-list state, but the builder owns thread-project fallback, project-execution tool classification, and timeline/tool-card enrichment so future relay or remote context metadata does not spread through model mutation code.
